+++ oneflow/core/eager/blob_instruction_type.cpp
@@ -130,5 +130,21 @@ void AccessBlobByCallbackInstructionType::Compute(vm::Instruction* instruction) 
   ptr->callback()(reinterpret_cast<uint64_t>(&ofblob));
 }
 
+class TouchTensorOnControlStreamInstructionType : public vm::InstructionType {
+ public:
+  TouchTensorOnControlStreamInstructionType() = default;
+  ~TouchTensorOnControlStreamInstructionType() = default;
+  using stream_type = vm::ControlStreamType;
+
+  void Infer(VirtualMachine* vm, Instruction* instruction) const override { UNIMPLEMENTED(); }
+  void Compute(VirtualMachine* vm, Instruction* instruction) const override {
+    // do nothing
+  }
+  void Infer(Instruction*) const override { UNIMPLEMENTED(); }
+  void Compute(Instruction*) const override { UNIMPLEMENTED(); }
+};
+COMMAND(vm::RegisterInstructionType<TouchTensorOnControlStreamInstructionType>(
+    "TouchTensorOnControlStream"));
+
 }  // namespace vm
 }  // namespace oneflow
