+++ oneflow/core/eager/cpu_blob_instruction_type.cpp
@@ -38,13 +38,5 @@ class CpuAccessBlobByCallbackInstructionType final : public AccessBlobByCallback
 COMMAND(vm::RegisterInstructionType<CpuAccessBlobByCallbackInstructionType>(
     "cpu.AccessBlobByCallback"));
 
-class CpuSoftSyncStreamInstructionType : public SoftSyncStreamInstructionType {
- public:
-  CpuSoftSyncStreamInstructionType() = default;
-  ~CpuSoftSyncStreamInstructionType() override = default;
-  using stream_type = vm::CpuStreamType;
-};
-COMMAND(vm::RegisterInstructionType<CpuSoftSyncStreamInstructionType>("cpu.SoftSyncStream"));
-
 }  // namespace vm
 }  // namespace oneflow
